refactor(opt): Remove forced type casting in cg from previous commit; refactor(solver): Remove forced type casting in PLBFGSAlg from previous commit; refactor(example,meshopt): Replace scipy's LinearOperator with FEALPy's LinearOperator; refactor(meshopt): Move opt, mesh_quality, radius_ratio_objective from mesh to meshopt.
