+++ fealpy/mesh/__init__.py
@@ -20,6 +20,3 @@ from .lagrange_triangle_mesh import LagrangeTriangleMesh
 from .lagrange_quadrangle_mesh import LagrangeQuadrangleMesh
 
 from .edge_mesh import EdgeMesh
-
-from .mesh_quality import RadiusRatioQuality
-from .radius_ratio_objective import RadiusRatioSumObjective
